Run `hi ...` command after all `nvim_set_hl` to fix intro screen
See https://github.com/rktjmp/lush.nvim/issues/94

Some kind of bug with `hi clear` and `nvim_set_hl` that wont set
highlights correctly on the intro screen. Applying any kind of highlight
repairs the colorscheme, so it's not exhibited when loading a file
directly as most filetype plugins are going to set a highlight. The
intro screen does not do this, so we mannually insert one `hi ...`
command after applying the colorscheme to ensure all highlights were
applied (or whatever is going on) until we can fix upstream.